# api/openapi.3.1.projection.yaml
...
openapi: 3.1.0
info:
  title: Bedrock API
  version: 2.0.0
  description: |
    Bedrock 2.0 HTTP API. Envelope `{ code, message, data? }`.
    Pagination data shape: `{ items, total, page, page_size, total_pages }`.
servers:
  - url: /api/v1
    description: Relative to the Bedrock host
security:
  - bearerAuth: []
...
paths:
...
  /roles:
    get:
      tags: [Roles]
      summary: List roles
      operationId: rolesList
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/RolePage"
        "403":
          $ref: "#/components/responses/Error"
    post:
      tags: [Roles]
      summary: Create role
      operationId: rolesCreate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Role"
        "400":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /roles/{id}:
    parameters:
      - $ref: "#/components/parameters/ID"
    get:
      tags: [Roles]
      summary: Get role
      operationId: rolesGet
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Role"
        "404":
          $ref: "#/components/responses/Error"
    put:
      tags: [Roles]
      summary: Update role
      operationId: rolesUpdate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Role"
        "400":
          $ref: "#/components/responses/Error"
    delete:
      tags: [Roles]
      summary: Delete role
      operationId: rolesDelete
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /roles/{id}/permissions:
    parameters:
      - $ref: "#/components/parameters/ID"
    put:
      tags: [Roles]
      summary: Replace role permission codes
      operationId: rolesSetPermissions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [permissions]
              properties:
                permissions:
                  type: array
                  items:
                    type: string
                    description: "{path}:action"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Role"
        "400":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /rbac/resources:
    get:
      tags: [RBAC]
      summary: List RBAC resource tree
      operationId: rbacResourcesList
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          items:
                            type: array
                            items:
                              $ref: "#/components/schemas/RbacResource"
        "403":
          $ref: "#/components/responses/Error"
    post:
      tags: [RBAC]
      summary: Create RBAC resource
      operationId: rbacResourcesCreate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RbacResourceCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/RbacResource"
        "400":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /rbac/resources/{id}:
    parameters:
      - $ref: "#/components/parameters/ID"
    get:
      tags: [RBAC]
      summary: Get RBAC resource
      operationId: rbacResourcesGet
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/RbacResource"
        "404":
          $ref: "#/components/responses/Error"
    put:
      tags: [RBAC]
      summary: Update RBAC resource
      operationId: rbacResourcesUpdate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RbacResourceUpdateRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/RbacResource"
        "400":
          $ref: "#/components/responses/Error"
    delete:
      tags: [RBAC]
      summary: Delete RBAC resource
      operationId: rbacResourcesDelete
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
        "400":
          $ref: "#/components/responses/Error"
  /menus:
    get:
      tags: [Menus]
      summary: List menu resource tree
      operationId: menusList
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          items:
                            type: array
                            items:
                              $ref: "#/components/schemas/RbacResource"
        "403":
          $ref: "#/components/responses/Error"
  /menus/{id}:
    parameters:
      - $ref: "#/components/parameters/ID"
    put:
      tags: [Menus]
      summary: Update menu metadata
      operationId: menusUpdate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RbacResourceUpdateRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/RbacResource"
        "400":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /menus/{id}/icon:
    parameters:
      - $ref: "#/components/parameters/ID"
    put:
      tags: [Menus]
      summary: Upload/replace level-1 menu icon (raw ≤32KB)
      operationId: menusUpdateIcon
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [icon_base64]
              properties:
                icon_base64:
                  type: string
                  description: Base64 or data URL; decoded raw size must be ≤32KB
                icon_mime:
                  type: string
                  example: image/png
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/RbacResource"
        "400":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    ID:
      name: id
      in: path
      required: true
      schema:
        type: integer
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    SuccessEnvelope:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
        message:
          type: string
        request_id:
          type: string
...
    Role:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        description:
          type: string
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/RolePermission"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    RolePermission:
      type: object
      properties:
        id:
          type: integer
        role_id:
          type: integer
        permission:
          type: string
    RoleCreateRequest:
      type: object
      required: [name, code]
      properties:
        name:
          type: string
        code:
          type: string
        description:
          type: string
        permissions:
          type: array
          items:
            type: string
    MenuMetadata:
      type: object
      properties:
        id:
          type: integer
        resource_id:
          type: integer
        title:
          type: string
        route:
          type: string
        icon_base64:
          type: string
        icon_mime:
          type: string
    RbacResource:
      type: object
      properties:
        id:
          type: integer
        path:
          type: string
        type:
          type: string
          enum: [menu, page, action, card]
        parent_id:
          type: integer
          nullable: true
        enabled:
          type: boolean
        sort_key:
          type: integer
        menu_metadata:
          $ref: "#/components/schemas/MenuMetadata"
        children:
          type: array
          items:
            $ref: "#/components/schemas/RbacResource"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    RbacResourceCreateRequest:
      type: object
      required: [path, type]
      properties:
        path:
          type: string
        type:
          type: string
          enum: [menu, page, action, card]
        parent_id:
          type: integer
          nullable: true
        enabled:
          type: boolean
        sort_key:
          type: integer
        title:
          type: string
        route:
          type: string
    RbacResourceUpdateRequest:
      type: object
      properties:
        enabled:
          type: boolean
        sort_key:
          type: integer
        title:
          type: string
        route:
          type: string
...
    Page:
      type: object
      required: [items, total, page, page_size, total_pages]
      properties:
        items:
          type: array
          items: {}
        total:
          type: integer
          format: int64
        page:
          type: integer
        page_size:
          type: integer
        total_pages:
          type: integer
...
    RolePage:
      allOf:
        - $ref: "#/components/schemas/Page"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Role"
...